python/requirements.txt
mysql-connector-python>=8.0.24
pyspark>=3.1.1
findspark>=1.4.2